backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.29.0
sqlalchemy>=2.0.3
alembic>=1.9.0
pydantic[email]>=1.9.0
psycopg2-binary>=2.9.5
asyncpg>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
h11>=0.14.0
